- Fixed structure of project - Used useEffect for saving todos to LS - Fixed anonim function to variable for getting initial todos value
